Fix warning: implicit conversion changes signedness
In file included from visp/modules/core/test/math/catchParticleFilter.cpp:49:
visp/modules/core/include/visp3/core/vpParticleFilter.h: In instantiation of ‘void vpParticleFilter<MeasurementsType>::updateMultithread(const MeasurementsType&) [with MeasurementsType = std::vector<vpImagePoint>]’:
visp/modules/core/include/visp3/core/vpParticleFilter.h:555:5:   required from ‘void vpParticleFilter<MeasurementsType>::update(const MeasurementsType&) [with MeasurementsType = std::vector<vpImagePoint>]’
visp/modules/core/include/visp3/core/vpParticleFilter.h:531:3:   required from ‘void vpParticleFilter<MeasurementsType>::filter(const MeasurementsType&, const double&, const vpColVector&) [with MeasurementsType = std::vector<vpImagePoint>]’
visp/modules/core/test/math/catchParticleFilter.cpp:555:22:   required from here
visp/modules/core/include/visp3/core/vpParticleFilter.h:734:14: warning: conversion to ‘unsigned int’ from ‘std::size_t’ {aka ‘long unsigned int’} may change the sign of the result [-Wsign-conversion]
  734 |     tempSums[static_cast<std::size_t>(iam)] = threadLikelihood<MeasurementsType>(m_likelihood, m_particles, z, m_w, istart, ipoints);
      |              ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~

+++ modules/core/include/visp3/core/vpParticleFilter.h
@@ -731,7 +731,7 @@ void vpParticleFilter<MeasurementsType>::updateMultithread(const MeasurementsTyp
       // last thread may do more
       ipoints = npoints - istart;
     }
-    tempSums[static_cast<std::size_t>(iam)] = threadLikelihood<MeasurementsType>(m_likelihood, m_particles, z, m_w, istart, ipoints);
+    tempSums[static_cast<unsigned int>(iam)] = threadLikelihood<MeasurementsType>(m_likelihood, m_particles, z, m_w, istart, ipoints);
   }
   sumWeights = tempSums.sum();
